Extract color helpers
