+++ src/index.ts
@@ -38,3 +38,4 @@ export * from './css/stylePresets';
 export * from './system/createFormStore';
 export * from './system/themeStore';
 export * from './hooks/useGoogleFonts';
+export * from './utilities/colors';
